+++ TombEngine/Game/Lara/lara_helpers.cpp
@@ -1661,14 +1661,14 @@ void SetLaraVehicle(ItemInfo* item, ItemInfo* vehicle)
 		if (previousVehicle == NO_VALUE)
 			return;
 
-		g_GameScript->OnVehicleExit(previousVehicle, false);
+		g_GameScript->OnVehicleLeave(previousVehicle, false);
 
 		auto vehicleObjectNumber = g_Level.Items[previousVehicle].ObjectNumber;
 		if (vehicleObjectNumber != ID_SPEEDBOAT && vehicleObjectNumber != ID_RUBBER_BOAT) // Leave boat vehicles active for inertia.
 			g_Level.Items[lara->Context.Vehicle].Active = false;
 
 		lara->Context.Vehicle = NO_VALUE;
-		g_GameScript->OnVehicleExit(previousVehicle, true);
+		g_GameScript->OnVehicleLeave(previousVehicle, true);
 	}
 	else
 	{