+++ TombEngine/Scripting/Include/ScriptInterfaceGame.h
@@ -63,7 +63,7 @@ public:
 	virtual void OnUseItem(short itemNumber, GAME_OBJECT_ID objectNumber) = 0;
 	virtual void OnPickup(short itemNumber, bool post) = 0;
 	virtual void OnVehicleEnter(short itemNumber, bool post) = 0;
-	virtual void OnVehicleExit(short itemNumber, bool post) = 0;
+	virtual void OnVehicleLeave(short itemNumber, bool post) = 0;
 	virtual void OnFreeze() = 0;
 
 	virtual void AddConsoleInput(const std::string& input) = 0;